api: Extract "args" module
